+++ Core/TigSubsystems/TigFont.cs
@@ -88,8 +88,8 @@ public class TigTextStyle
      */
 public struct TigFontMetrics
 {
-    public int width;
-    public int height;
+    public float width;
+    public float height;
     public int lines;
     public int lineheight;
 }
